fix: update serverless config and fix dynamodb insertion issues
